lumos/backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
mysql-connector-python==8.0.33
pydantic==1.10.7
python-jose==3.3.0
passlib==1.7.4
requests==2.31.0
openai==1.75.0
celery[redis]==5.2.7
httpx
psutil==5.9.5
jinja2==3.1.2